Add function to set event listeners on each square
The function perpasses the whole array of squares and
sets an event listener on each one, modifying its color
when triggered.

This is important for scalability as well, as we aren't
inserting listeners on each element manually. Although,
thinking about scalability, it would be interesting to
consider inserting only one listener that considers the
target, because it would consume less memory.