Add button to set number of squares per row
The function called when the button is clicked
deletes all rows to create new ones using the
parameter passed by the user. As the function
that generates the grid actually just append
squares on each row, generating it again without
removing previous ones would cause it to exceed
the amount intended.

The potential big amount of event handlers (one
per square) can cause the application to be slow.
So, it's recommended to adjust it, on future updates,
to only one event handler on a parent node that
accounts target node.