Add fading-in opacity on drawing action
There's low opacity on the first elements painted, that grows
at a 0.01% rate at each square painted. However, the fading in
isn't reset on each user's interaction with the board, and it
would be an interesting feature for a future version.

It could be achieved with an event handler on the whole board.