Reset fading-in when mouse leaves board
It creates the effect of increased strength
on the hand that is drawing, for each "turn".

A nice feature would be to set the drawing
only when mouse is clicking, and reset the fading
when stop clicking.

Also, a button to set single color, color value
or colored randomly would be interesting.